+++ srf_generation/source_parameter_generation/gcmt_to_realisation.py
@@ -2,19 +2,13 @@
 
 import argparse
 from logging import Logger
-from multiprocessing import pool
 from os import makedirs
 from os.path import abspath, isfile, dirname, join
 from typing import Callable, Union, Dict, Any
 
 import pandas as pd
 
-from qcore.formats import load_fault_selection_file
-from qcore.simulation_structure import (
-    get_realisation_name,
-    get_srf_path,
-    get_sources_dir,
-)
+from qcore.simulation_structure import get_realisation_name
 from qcore.qclogging import (
     get_logger,
     add_general_file_handler,
@@ -42,6 +36,21 @@ GCMT_FILE_COLUMNS = [
 ]
 
 UNPERTURBATED = "unperturbated"
+GCMT_DTYPE = {
+    GCMT_FILE_COLUMNS[i]: col_type
+    for i, col_type in enumerate(
+        [
+            "object",
+            "float64",
+            "float64",
+            "float64",
+            "float64",
+            "float64",
+            "float64",
+            "float64",
+        ]
+    )
+}
 
 
 def load_args(primary_logger: Logger):
@@ -49,18 +58,13 @@ def load_args(primary_logger: Logger):
         "Generates realisation files from gcmt and other input data."
     )
 
-    parser.add_argument("fault_selection_file", type=abspath)
+    parser.add_argument("fault_name")
+    parser.add_argument("realisation_count", type=int)
     parser.add_argument("gcmt_file", type=abspath)
     parser.add_argument("--version", type=str, default="unperturbated")
-    parser.add_argument(
-        "-n",
-        "--n_processes",
-        default=1,
-        type=int,
-        help="Number of processes to generate realisations with. "
-        "Setting this higher than the number of faults will not have any additional effect.",
-    )
-    parser.add_argument("-c", "--cybershake_root", type=abspath, default=abspath("."))
+    parser.add_argument("--vel_mod_1d", type=abspath)
+    parser.add_argument("--vel_mod_1d_out", type=abspath)
+    parser.add_argument("-o", "--output_dir", type=abspath, default=abspath("."))
     parser.add_argument(
         "-a",
         "--aggregate_file",
@@ -86,13 +90,28 @@ def load_args(primary_logger: Logger):
     args = parser.parse_args()
 
     errors = []
-    if not isfile(args.fault_selection_file):
-        errors.append(f"Specified station file not found: {args.fault_selection_file}")
+
     if not isfile(args.gcmt_file):
         errors.append(f"Specified gcmt file not found: {args.gcmt_file}")
+
     if args.aggregate_file is not None and isfile(args.aggregate_file):
         errors.append(
             f"Specified aggregation file {args.aggregate_file} already exists, please choose another file"
+        )
+    if args.vel_mod_1d is not None and not isfile(args.vel_mod_1d):
+        errors.append(
+            f"Specified 1d velocity model file {args.vel_mod_1d} does not exist"
+        )
+    if (args.vel_mod_1d is not None and args.vel_mod_1d_out is None) or (
+        args.vel_mod_1d is not None and args.vel_mod_1d_out is None
+    ):
+        if args.vel_mod_1d is None:
+            given = "vel_mod_1d_out"
+        else:
+            given = "vel_mod_1d"
+        errors.append(
+            f"If either of the arguments vel_mod_1d and vel_mod_1d_out are given the other must also be given. "
+            f"{given} was given"
         )
 
     for i, (param_name, filepath) in enumerate(args.source_parameter):
@@ -112,9 +131,9 @@ def load_args(primary_logger: Logger):
         )
         raise ValueError(message)
 
-    makedirs(args.cybershake_root, exist_ok=True)
+    makedirs(args.output_dir, exist_ok=True)
 
-    log_file = join(args.cybershake_root, primary_logger.name + "_log.txt")
+    log_file = join(args.output_dir, primary_logger.name + "_log.txt")
     add_general_file_handler(primary_logger, log_file)
 
     primary_logger.debug(f"Arguments parsed: {args}")
@@ -122,12 +141,36 @@ def load_args(primary_logger: Logger):
     return args
 
 
-def generate_uncertainties(
+def load_1d_velocity_mod(vel_mod_1d):
+    return pd.read_csv(
+        vel_mod_1d,
+        delim_whitespace=True,
+        header=None,
+        skiprows=1,
+        names=["depth", "vp", "vs", "rho", "qp", "qs"],
+    )
+
+
+def save_1d_velocity_model(
+    perturbed_realisation: pd.DataFrame, vel_mod_1d_dir, realisation_name
+):
+    file_name = join(vel_mod_1d_dir, f"{realisation_name}.v1d")
+    with open(file_name, "w") as v1d:
+        v1d.write(f"{len(perturbed_realisation)}\n")
+        for line in perturbed_realisation.itertuples():
+            v1d.write(
+                f"{line.depth} {line.vp} {line.vs} {line.rho} {line.qp} {line.qs}\n"
+            )
+
+
+def generate_fault_realisations(
     data: GCMT_Source,
     realisation_count: int,
-    cybershake_root: str,
+    output_directory: str,
     perturbation_function: Callable,
     aggregate_file: Union[str, None],
+    vel_mod_1d: pd.DataFrame,
+    vel_mod_1d_dir: str,
     primary_logger_name: str,
     additional_source_parameters: Dict[str, Any],
 ):
@@ -137,40 +180,21 @@ def generate_uncertainties(
     fault_name = data.pid
 
     for i in range(1, realisation_count + 1):
+        realisation_name = get_realisation_name(fault_name, i)
+        realisation_file_name = join(output_directory, f"{realisation_name}.csv")
         fault_logger.debug(
-            f"Generating realisation {i} of {realisation_count} for fault {data.pid}"
+            f"Generating realisation {i} of {realisation_count} for fault {fault_name}"
         )
-
-        fault_logger.debug("Calling perturbation function.")
-
-        perturbed_realisation = perturbation_function(
-            sources_line=data, additional_source_parameters=additional_source_parameters
-        )
-
-        fault_logger.debug(
-            f"Got results from perturbation_function: {perturbed_realisation}"
-        )
-        perturbed_realisation["name"] = get_realisation_name(fault_name, i)
-
-        file_name = get_srf_path(
-            cybershake_root, perturbed_realisation["name"]
-        ).replace(".srf", ".csv")
-        makedirs(dirname(file_name), exist_ok=True)
-        fault_logger.debug(
-            f"Created Srf directory and attempting to save perturbated source generation parameters there: {file_name}"
-        )
-        rel_df = pd.DataFrame(perturbed_realisation, index=[i])
-
-        rel_df.to_csv(file_name, index=False)
-
-        if aggregate_file is not None:
-            if not isfile(aggregate_file):
-                rel_df.to_csv(aggregate_file)
-            else:
-                rel_df.to_csv(aggregate_file, mode="a", header=False)
-
-        fault_logger.debug(
-            f"Parameters saved succesfully. Continuing to next realisation if one exists."
+        generate_realisation(
+            realisation_file_name,
+            realisation_name,
+            perturbation_function,
+            data,
+            additional_source_parameters,
+            aggregate_file,
+            vel_mod_1d,
+            vel_mod_1d_dir,
+            fault_logger,
         )
 
     unperturbated_function = load_perturbation_function(UNPERTURBATED)
@@ -179,10 +203,51 @@ def generate_uncertainties(
             sources_line=data, additional_source_parameters=additional_source_parameters
         )
         rel_df = pd.DataFrame(unperturbated_realisation, index=[0])
-        file_name = join(
-            get_sources_dir(cybershake_root), fault_name, f"{fault_name}.csv"
-        )
-        rel_df.to_csv(file_name, index=False)
+        realisation_file_name = join(output_directory, f"{fault_name}.csv")
+        rel_df.to_csv(realisation_file_name, index=False)
+
+
+def generate_realisation(
+    realisation_file_name,
+    realisation_name,
+    perturbation_function,
+    data,
+    additional_source_parameters,
+    aggregate_file,
+    vel_mod_1d,
+    vel_mod_1d_dir,
+    fault_logger,
+):
+    fault_logger.debug("Calling perturbation function.")
+    perturbed_realisation = perturbation_function(
+        sources_line=data,
+        additional_source_parameters=additional_source_parameters,
+        vel_mod_1d=vel_mod_1d,
+    )
+    fault_logger.debug(
+        f"Got results from perturbation_function: {perturbed_realisation}"
+    )
+    perturbed_realisation["params"]["name"] = realisation_name
+    makedirs(dirname(realisation_file_name), exist_ok=True)
+    fault_logger.debug(
+        f"Created Srf directory and attempting to save perturbated source generation parameters there: {realisation_file_name}"
+    )
+    rel_df = pd.DataFrame(perturbed_realisation["params"], index=[0])
+    rel_df.to_csv(realisation_file_name, index=False)
+
+    if aggregate_file is not None:
+        if not isfile(aggregate_file):
+            rel_df.to_csv(aggregate_file)
+        else:
+            rel_df.to_csv(aggregate_file, mode="a", header=False)
+
+    if vel_mod_1d is not None and vel_mod_1d_dir is not None:
+        perturbed_vel_mod_1d = perturbed_realisation["vel_mod_1d"]
+        save_1d_velocity_model(perturbed_vel_mod_1d, vel_mod_1d_dir, realisation_name)
+
+    fault_logger.debug(
+        f"Parameters saved succesfully. Continuing to next realisation if one exists."
+    )
 
 
 def main():
@@ -194,46 +259,25 @@ def main():
     perturbation_function = load_perturbation_function(args.version)
     primary_logger.debug(f"Perturbation function loaded. Version: {args.version}")
 
-    faults = load_fault_selection_file(args.fault_selection_file)
-    types = [
-        "object",
-        "float64",
-        "float64",
-        "float64",
-        "float64",
-        "float64",
-        "float64",
-        "float64",
+    gcmt_data = pd.read_csv(args.gcmt_file, usecols=GCMT_FILE_COLUMNS, index_col=0)[
+        ["Latitude", "Longitude", "CD", "Mw", "strike1", "dip1", "rake1"]
     ]
-    dtype = {GCMT_FILE_COLUMNS[i]: types[i] for i in range(len(types))}
-
-    gcmt_data = pd.read_csv(args.gcmt_file, usecols=GCMT_FILE_COLUMNS, dtype=dtype)[
-        ["PublicID", "Latitude", "Longitude", "CD", "Mw", "strike1", "dip1", "rake1"]
-    ]
-    gcmt_data.columns = GCMT_PARAM_NAMES
+    gcmt_data.columns = GCMT_PARAM_NAMES[1:]
 
     primary_logger.debug(
-        f"{len(faults.keys())} faults were selected and gcmt data for {len(gcmt_data)} faults were loaded. "
-        f"Running cross comparison now."
+        f"GCMT data for {len(gcmt_data)} faults were loaded. Checking for fault {args.fault_name}"
     )
 
-    gcmt_lines = list(
-        gcmt_data[gcmt_data["pid"].isin(faults.keys())].itertuples(
-            name=None, index=False
+    if args.fault_name in gcmt_data.index:
+        gcmt_line = gcmt_data.loc[args.fault_name]
+    else:
+        message = (
+            f"Fault {args.fault_name} is not available in the given GCMT file, exiting"
         )
-    )
-
-    pids = [gcmt[0] for gcmt in gcmt_lines]
-    missing_faults = [fault for fault in faults.keys() if fault not in pids]
-
-    if missing_faults:
-        message = f"Some fault(s) in the fault selection were not found in the gcmt file: {missing_faults}"
         primary_logger.log(NOPRINTCRITICAL, message)
+        primary_logger.debug("Available faults are as follows:")
+        primary_logger.debug(gcmt_data.index)
         raise ValueError(message)
-
-    primary_logger.debug(
-        f"All {len(gcmt_lines)} faults specified in the fault selection file were found"
-    )
 
     additional_source_parameters = pd.DataFrame()
 
@@ -250,64 +294,41 @@ def main():
             parameter_df, how="outer"
         )
 
-    messages = generate_messages(
-        additional_source_parameters,
-        args.aggregate_file,
-        args.cybershake_root,
-        faults,
-        gcmt_lines,
-        perturbation_function,
-        primary_logger,
-    )
+    additional_source_specific_data = {}
+    if gcmt_line[0] in additional_source_parameters.index:
+        additional_source_specific_data = additional_source_parameters.loc[
+            gcmt_line[0]
+        ].to_dict()
 
-    if len(messages) < args.n_processes:
-        n_processes = len(messages)
-        primary_logger.info(
-            f"The number of processes requested to create realisations ({args.n_processes}) is greater than "
-            f"the number of faults ({n_processes}), this value will be capped to the number of faults"
+    if args.vel_mod_1d is not None:
+        vel_mod_1d_layers = load_1d_velocity_mod(args.vel_mod_1d)
+    else:
+        vel_mod_1d_layers = None
+
+    if args.realisation_count > 1:
+        generate_fault_realisations(
+            GCMT_Source(args.fault_name, *gcmt_line),
+            args.realisation_count,
+            args.output_dir,
+            perturbation_function,
+            args.aggregate_file,
+            vel_mod_1d_layers,
+            args.vel_mod_1d_out,
+            primary_logger.name,
+            additional_source_specific_data,
         )
     else:
-        n_processes = args.n_processes
-
-    primary_logger.debug(
-        f"{len(messages)} messages were created to create a total of {sum([m[1] for m in messages])} realisations"
-    )
-
-    worker_pool = pool.Pool(processes=n_processes)
-    worker_pool.starmap(generate_uncertainties, messages)
-
-
-def generate_messages(
-    additional_source_parameters: pd.DataFrame,
-    aggregate_file,
-    cybershake_root,
-    faults,
-    gcmt_lines,
-    perturbation_function,
-    primary_logger,
-):
-    messages = []
-    for i in range(len(faults)):
-        gcmt_data = gcmt_lines[i]
-
-        additional_source_specific_data = {}
-        if gcmt_data[0] in additional_source_parameters.index:
-            additional_source_specific_data = additional_source_parameters.loc[
-                gcmt_data[0]
-            ].to_dict()
-
-        messages.append(
-            (
-                GCMT_Source(*gcmt_data),
-                faults[gcmt_data[0]],
-                cybershake_root,
-                perturbation_function,
-                aggregate_file,
-                primary_logger.name,
-                additional_source_specific_data,
-            )
+        generate_realisation(
+            join(args.output_dir, f"{args.fault_name}.csv"),
+            args.fault_name,
+            perturbation_function,
+            GCMT_Source(args.fault_name, *gcmt_line),
+            additional_source_parameters,
+            args.aggregate_file,
+            args.vel_mod_1d,
+            args.vel_mod_1d_out,
+            primary_logger,
         )
-    return messages
 
 
 if __name__ == "__main__":