+++ srf_generation/source_parameter_generation/uncertainties/versions/unperturbated.py
@@ -2,7 +2,7 @@
 from typing import Any, Dict, Union
 
 from srf_generation.source_parameter_generation.uncertainties.common import (
-    verify_params,
+    verify_realisation_params,
     GCMT_Source,
     NHM_Source,
 )
@@ -11,6 +11,7 @@ from srf_generation.source_parameter_generation.uncertainties.common import (
 def generate_source_params(
     sources_line: Union[GCMT_Source, NHM_Source],
     additional_source_parameters: Dict[str, Any],
+    **kwargs
 ) -> Dict[str, Any]:
     """source_data should have the following parameters available via . notation:
       - source_data.pid: name of the event
@@ -23,7 +24,9 @@ def generate_source_params(
       - source_data.rake
     """
 
-    params = {
+    realisation = kwargs
+
+    realisation["params"] = {
         "type": 1,
         "name": sources_line.pid,
         "latitude": sources_line.lat,
@@ -34,7 +37,7 @@ def generate_source_params(
         "dip": sources_line.dip,
         "rake": sources_line.rake,
     }
-    params.update(additional_source_parameters)
+    realisation["params"].update(additional_source_parameters)
 
-    verify_params(params)
-    return params
+    verify_realisation_params(realisation["params"])
+    return realisation